# my-blog-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  my-blog-backend_internal_api_v1_dto_request.EmailCaptchaRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      username:
        description: 用户名可选，如果未提供则使用邮箱
        example: admin
        type: string
    required:
    - email
    type: object
  my-blog-backend_internal_api_v1_dto_request.LoginByEmailRequest:
    properties:
      code:
        maxLength: 50
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  my-blog-backend_internal_api_v1_dto_request.LoginRequest:
    properties:
      captchaAnswer:
        type: string
      captchaId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captchaAnswer
    - captchaId
    - password
    - username
    type: object
  my-blog-backend_internal_api_v1_dto_request.RegisterRequest:
    properties:
      code:
        maxLength: 50
        type: string
      email:
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - code
    - email
    - password
    - username
    type: object
...
  my-blog-backend_internal_api_v1_dto_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@myblog.com
    name: API Support
    url: http://www.myblog.com/support
  description: This is a blog backend API with Swagger documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: My Blog API
  version: "1.0"
paths:
...
  /api/v1/auth/captcha:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 验证码类型
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取验证码
      tags:
      - 认证管理
  /api/v1/auth/email/code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.EmailCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 发送邮箱验证码
      tags:
      - 认证管理
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 用户名密码登录
      tags:
      - 认证管理
  /api/v1/auth/login/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.LoginByEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 邮箱验证码登录
      tags:
      - 认证管理
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: 令牌信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 刷新令牌
      tags:
      - 认证管理
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 用户注册
      tags:
      - 认证管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"